Move util functions into their own module
